recorder: Change track frame rate to floating point
- The track frame rate can be configured as float by client.
- gtest is added as well.

+++ include/qmmf-sdk/qmmf_recorder_params.h
@@ -175,7 +175,7 @@ struct VideoTrackCreateParam {
   uint32_t         camera_id;
   uint32_t         width;
   uint32_t         height;
-  uint32_t         frame_rate;
+  float            frame_rate;
   VideoFormat      format_type;
   VideoCodecParams codec_param;
   uint32_t         out_device;
